+++ tests/genome/test_prodigal.py
@@ -2,15 +2,16 @@
 """
  * @Date: 2022-10-11 20:11:35
  * @LastEditors: Hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2023-12-21 20:09:59
+ * @LastEditTime: 2023-12-22 14:57:23
  * @FilePath: /genome/tests/genome/test_prodigal.py
  * @Description:
 __file__ = "test/genome/test_prodigal.py"
 """
 
 from pathlib import Path
-import shutil
-from genome.prodigal import prodigal_gff_onethread
+from Bio import SeqIO
+from genome.bin_statistic_ext import format_bin_input
+from genome.prodigal import prodigal_gff_onethread, prodigal_multithread
 
 try:
     from _decorator import temp_output, test_temp, test_files
@@ -21,38 +22,35 @@ except (ModuleNotFoundError, ImportError):
 @temp_output
 def test_prodigal_gff_onethread(test_temp: Path):
     genome = test_files / "binny_contigs_4bins.fa"
-    expect = test_files / "binny_contigs_4bins-prodigal.single.gff"
 
-    test_out = test_temp / "binny_contigs_4bins-prodigal.single.gff"
-    gff = prodigal_gff_onethread(shutil.copy(genome, test_temp), "single")
-    assert gff == test_out.expanduser().absolute()
+    SeqIO.write(
+        (i for i, _ in zip(SeqIO.parse(genome, "fasta"), range(10))),
+        test_out_fa := test_temp / "binny_contigs_4bins-top10.fa",
+        "fasta",
+    )
+    expect = test_files / "binny_contigs_4bins-top10-prodigal.gvmeta.gff"
+
+    gff = prodigal_gff_onethread(test_out_fa, "gvmeta")
+    assert gff == (test_temp / expect.name).expanduser().absolute()
 
     assert gff.read_text() == expect.read_text()
 
 
 @temp_output
-def test_pyrodigal(test_temp: Path):
-    genome = test_files / "binny_contigs_4bins.fa"
-    from Bio import SeqIO
-    import pyrodigal
-    import pyrodigal_gv
-
-    seqs = SeqIO.to_dict(SeqIO.parse(genome, "fasta"))
-
-    gf = pyrodigal.GeneFinder(meta=True)
-
-    gfv = pyrodigal_gv.ViralGeneFinder(meta=True)
-
-    with open("a", "w") as fa, open("b", "w") as fb:
-        for i in seqs.values():
-            gf.find_genes(bytes(i.seq)).write_gff(
-                fa, i.id, include_translation_table=True
-            )
-            gfv.find_genes(bytes(i.seq)).write_gff(
-                fb, i.id, include_translation_table=True
-            )
-    # gene1 = genes[0]
-    # gene1.begin
-
-    gf = pyrodigal.GeneFinder(meta=False)
-    gf.train(*(bytes(i.seq) for i in seqs.values()))
+def test_prodigal_multithread(test_temp: Path):
+    ctg2mag = test_files / "binny_unitem_unanimous.tsv"
+    support = test_files / "binny_contigs_4bins.fa"
+    bin_input_dir, binids, suffix = format_bin_input(
+        bin_output=test_temp / "prodigal",
+        bin_input=ctg2mag,
+        support=support,
+        keep_if_avail=False,
+    )
+    for bin_faa in prodigal_multithread(
+        (bin_input_dir / (binid + suffix) for binid in binids),
+        mode="single",
+        out_dir=bin_input_dir,
+        suffix="-ge33.faa",
+        threads=2,
+    ):
+        bin_faa.rename(str(bin_faa)[:-25] + ".faa")